# specs/api.yaml
openapi: 3.1.0
info:
  title: PsiZero Resonance Platform API
  version: 1.0.0
  description: >
    Integrated APIs for symbolic resonance technologies:
    - SRS (Symbolic Resonance Solver) for NP-complete problems
    - HQE (Holographic Quantum Encoder) simulator
    - QSEM (Quantum Semantics) for concept resonance
    - NLC (Non-Local Communication) prime-eigenstate channels
    - QCR (Quantum Consciousness Resonator) sessions
    - I-CHING oracle (hexagram evolution)
    - Unified resonance physics (observer-capacity/gravity model)
servers:
  - url: https://api.psizero.dev
    description: Production
  - url: https://sandbox.psizero.dev
    description: Sandbox
...
security:
  - ApiKeyAuth: []
  - OAuth2CC: [platform.read, platform.write]
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    OAuth2CC:
      type: oauth2
      description: Client Credentials flow (for enterprises)
      flows:
        clientCredentials:
          tokenUrl: https://auth.psizero.dev/oauth/token
          scopes:
            platform.read: Read access
            platform.write: Write/access compute
...
  responses:
    ProblemJson:
      description: RFC 7807 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        meta: { type: object, additionalProperties: true }
...
    IChingEvolveRequest:
      type: object
      properties:
        question: { type: string }
        steps: { type: integer, default: 7 }
      required: [question]
    IChingStep:
      type: object
      properties:
        step: { type: integer }
        hexagram: { type: string, pattern: '^[01]{6}$', description: top->bottom lines }
        entropy: { type: number }
        attractorProximity: { type: number }
    IChingEvolveResponse:
      type: object
      properties:
        sequence:
          type: array
          items: { $ref: '#/components/schemas/IChingStep' }
        stabilized: { type: boolean }

    # ---------- Unified / Gravity ----------
...
paths:
...
  /v1/iching/evolve:
    post:
      tags: [IChing]
      summary: Evolve a hexagram sequence from a question
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IChingEvolveRequest' }
      responses:
        '200':
          description: Sequence and stabilization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IChingEvolveResponse' }
        '400': { $ref: '#/components/responses/ProblemJson' }

  # ---------- Unified ----------
...